requirements/base.txt
# SHA1:b79eb5fdcba5c262a8f0ff2f7c73a6b0e3133b8d
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
asgiref==3.8.1
    # via django
beautifulsoup4==4.12.3
    # via -r requirements/base.in
certifi==2024.12.14
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via
    #   reportlab
    #   requests
cryptography==44.0.0
    # via
    #   josepy
    #   mozilla-django-oidc
    #   pyopenssl
diff-match-patch==20241021
    # via django-import-export
django==4.2.17
    # via
    #   -r requirements/base.in
    #   django-guardian
    #   django-import-export
    #   django-simple-history
    #   djangorestframework
    #   mozilla-django-oidc
django-adminactions==1.9
    # via -r requirements/base.in
django-better-admin-arrayfield==1.4.2
    # via -r requirements/base.in
django-guardian==2.4.0
    # via -r requirements/base.in
django-import-export==4.3.9
    # via -r requirements/base.in
django-modelclone @ git+https://github.com/al-beton/django-modelclone/
    # via -r requirements/base.in
django-simple-history==3.7.0
    # via -r requirements/base.in
django4-background-tasks==1.2.10
    # via -r requirements/base.in
djangoql==0.17.1
    # via -r requirements/base.in
djangorestframework==3.16.0
    # via -r requirements/base.in
et-xmlfile==2.0.0
    # via openpyxl
idna==3.10
    # via requests
josepy==1.14.0
    # via mozilla-django-oidc
lxml==4.9.0
    # via -r requirements/base.in
mozilla-django-oidc==4.0.1
    # via -r requirements/base.in
openpyxl==3.1.2
    # via -r requirements/base.in
pillow==11.1.0
    # via
    #   -r requirements/base.in
    #   reportlab
ply==3.11
    # via djangoql
psycopg==3.2.3
    # via -r requirements/base.in
pycparser==2.22
    # via cffi
pyopenssl==24.3.0
    # via josepy
pytz==2024.2
    # via django-adminactions
pyzmq==26.2.0
    # via zmq
reportlab==4.4.3
    # via -r requirements/base.in
requests==2.32.3
    # via mozilla-django-oidc
six==1.17.0
    # via django4-background-tasks
soupsieve==2.6
    # via beautifulsoup4
sqlparse==0.5.3
    # via django
tablib==3.7.0
    # via django-import-export
typing-extensions==4.12.2
    # via psycopg
urllib3==2.3.0
    # via requests
xlrd==1.1.0
    # via
    #   -r requirements/base.in
    #   django-adminactions
xlwt==1.3.0
    # via django-adminactions
zmq==0.0.0
    # via -r requirements/base.in
